+++ lib/src/widgets/adaptive_scaffold.dart
@@ -1,6 +1,7 @@
 import 'package:adaptive_platform_ui/src/widgets/ios26/ios26_native_tab_bar.dart';
 import 'package:flutter/cupertino.dart';
 import 'package:flutter/material.dart';
+import 'package:flutter/services.dart';
 import '../platform/platform_info.dart';
 import '../style/sf_symbol.dart';
 import 'adaptive_app_bar.dart';
@@ -521,6 +522,8 @@ class _AdaptiveScaffoldState extends State<AdaptiveScaffold> {
             unselectedItemColor:
                 widget.bottomNavigationBar!.unselectedItemColor,
             hidden: widget.tabBarHidden,
+            gestureBlockingPolicy:
+                widget.bottomNavigationBar!.gestureBlockingPolicy,
           );
         }
         // iOS 26+ with useNativeBottomBar=false OR iOS <26
@@ -1007,6 +1010,8 @@ class _MinimizableTabBar extends StatefulWidget {
     this.selectedItemColor,
     this.unselectedItemColor,
     this.hidden = false,
+    this.gestureBlockingPolicy =
+        UiKitViewGestureBlockingPolicy.fallbackToPluginDefault,
   });
 
   final int selectedIndex;
@@ -1017,6 +1022,7 @@ class _MinimizableTabBar extends StatefulWidget {
   final Color? selectedItemColor;
   final Color? unselectedItemColor;
   final bool hidden;
+  final UiKitViewGestureBlockingPolicy gestureBlockingPolicy;
 
   @override
   State<_MinimizableTabBar> createState() => _MinimizableTabBarState();
@@ -1127,6 +1133,7 @@ class _MinimizableTabBarState extends State<_MinimizableTabBar>
         unselectedItemTint: widget.unselectedItemColor,
         minimizeBehavior: widget.minimizeBehavior,
         hidden: widget.hidden,
+        gestureBlockingPolicy: widget.gestureBlockingPolicy,
       ),
     );
   }